+++ Commands/ExperienceCommands.cs
@@ -36,7 +36,7 @@ public static class ExperienceCommands {
         ctx.Reply(response);
     }
 
-    [Command("set", "s", "<playerName> <level>", "Sets the specified player's level to the start of the given level", adminOnly: true)]
+    [Command("set", "s", "<playerName> <level>", "Sets the specified player's level to the start of the given level", adminOnly: false)]
     public static void SetLevel(ChatCommandContext ctx, string name, int level){
         if (!Plugin.ExperienceSystemActive){
             ctx.Reply("Experience system is not enabled.");
@@ -165,24 +165,39 @@ public static class ExperienceCommands {
         ctx.Reply("Ability level up points reset.");
     }
     
-    [Command(name: "bump20", shortHand: "", adminOnly: false, usage: "[PlayerName]", description: "Temporarily bumps the player to lvl20 so they can skip the quest")]
-    public static void BumpToLevel20(ChatCommandContext ctx, string playerName = null)
+    [Command(name: "bump20", shortHand: "", adminOnly: false, usage: "", description: "Temporarily bumps you to lvl20 so you can skip the quest")]
+    public static void BumpToLevel20(ChatCommandContext ctx)
     {
         var playerEntity = ctx.Event.SenderCharacterEntity;
         var userEntity = ctx.Event.SenderUserEntity;
 
-        if (playerName != null)
-        {
-            if (!PlayerCache.FindPlayer(playerName, true, out playerEntity, out userEntity))
-            {
-                throw ctx.Error("Player not found.");
-            }
-        }
-
         try
         {
             SetUserLevel(ctx.Event.SenderCharacterEntity, ctx.Event.SenderUserEntity, ctx.User.PlatformId, 20, 5);
-            ctx.Reply($"Player has been bumped to lvl 20 for 5 seconds. Equip an item and then claim the reward.");
+            ctx.Reply($"You has been bumped to lvl 20 for 5 seconds. Equip an item and then claim the reward.");
+        }
+        catch (Exception e)
+        {
+            Plugin.Log(Plugin.LogSystem.Core, LogLevel.Error, $"Failed to bump20 {e.Message}", true);
+            throw ctx.Error($"Failed to bump20. Check logs for more details.");
+        }
+    }
+    
+    [Command(name: "bump20", shortHand: "", adminOnly: false, usage: "<PlayerName>", description: "Temporarily bumps the player to lvl20 so they can skip the quest")]
+    public static void BumpToLevel20(ChatCommandContext ctx, string playerName)
+    {
+        if (!PlayerCache.FindPlayer(playerName, true, out var playerEntity, out var userEntity))
+        {
+            throw ctx.Error("Player not found.");
+        }
+
+        var steamId = PlayerCache.GetSteamIDFromName(playerName);
+
+        try
+        {
+            SetUserLevel(playerEntity, userEntity, steamId, 20, 5);
+            ctx.Reply($"Player has been bumped to lvl 20 for 5 seconds.");
+            Output.SendMessage(userEntity, "You have been bumped to lvl 20 for 5 seconds. Equip an item and then claim the reward.");
         }
         catch (Exception e)
         {